reverted i2cTest to old. Added new file for testing instead
